+++ src/header/util/error.h
@@ -22,66 +22,16 @@ namespace minichess_AI
 
         // manipulate variables
 
-        std::string GetMessage()
-        {
-            return message;
-        }
+        std::string GetMessage();
 
         std::string DisplayError();
 
         // operator
 
-        // assignment operator
-        MCError operator=(const MCError &mce)
-        {
-            status = mce.status;
-            message = mce.message;
-            return *this;
-        }
-
-        // same status
-        bool operator==(const MCError &mce)
-        {
-            if (status == mce.status)
-                return true;
-            else
-                return false;
-        }
-
-        // not same status
-        bool operator!=(const MCError &mce)
-        {
-            if (status != mce.status)
-                return true;
-            else
-                return false;
-        }
+        MCError operator=(const MCError &);
+        bool operator==(const MCError &);
+        bool operator!=(const MCError &);
     };
-
-    // definitions
-
-    // get error description as std::string
-    // return (error type + error message)
-    std::string MCError::DisplayError()
-    {
-        switch (status)
-        {
-        case 0:
-            return "";
-        case 100:
-            return "FEN Error : " + message;
-        case 200:
-            return "Move Error : " + message;
-        case 300:
-            return "Move Wrong Piece Error : " + message;
-        case 10000:
-            return "Test Error : " + message;
-        default:
-            return "Unknown Error : " + message;
-        }
-
-        return "";
-    }
 
     // error types
