+++ src/header/util/util.h
@@ -1,5 +1,8 @@
 #pragma once
 
-#define max(a, b) (((a) > (b)) ? (a) : (b))
-#define min(a, b) (((a) < (b)) ? (a) : (b))
-#define abs(a) ((a >= 0) ? (a) : -(a))
+template <typename T>
+T max(T a, T b) { return (a > b) ? a : b; };
+template <typename T>
+T min(T a, T b) { return (a < b) ? a : b; };
+template <typename T>
+T abs(T a) { return a > 0 ? a : -a; };
